Split parser functions out into Require.Parser
In the same step the types into which the file contents are parsed live
now in Require.Types as they are not specific to the parser.
